% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/040_include_git_ignore.R
\name{copy_git_ignore}
\alias{copy_git_ignore}
\title{Create .gitignore to keep files out of github repository}
\usage{
copy_git_ignore()
}
\arguments{
\item{license}{may be "mit"}
}
\value{
none
}
\description{
This function omits files from a github repository a license to the current repository
}
\keyword{.gitignore}
\keyword{github}
