% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020_affix_mit_license.R
\name{copy_mit_license}
\alias{copy_mit_license}
\title{Add license to current directory}
\usage{
copy_mit_license(license = "mit")
}
\arguments{
\item{license}{may be "mit"}
}
\value{

}
\description{
This function adds an MIT license to the current repository
}
\examples{
copy_mit_license("mit")
#' @section Warning:
}
\keyword{add}
\keyword{license}
