% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/030_copy_readme.R
\name{copy_readme}
\alias{copy_readme}
\title{Create README for push to github account.}
\usage{
copy_readme()
}
\arguments{
\item{path}{defaults to working directory}
}
\description{
Creates a README template for use in github account
}
\examples{
copy_readme()

}
\keyword{README}
\keyword{github}
\keyword{template}
