% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010_create_folders.R
\name{create_folders}
\alias{create_folders}
\title{Create skeleton folder tree for data projects.}
\usage{
create_folders(path = getwd())
}
\arguments{
\item{path}{defaults to working directory}
}
\description{
Creates a set of folders for a basic data project. Folders created
include:  data/tidy, data/pure, docs, R, figs, tbls
}
\examples{
create_folders()

}
\keyword{directory}
\keyword{folder}
\keyword{tree}
