% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050_insert_readme_into_every_dir.R
\name{insert_readme_into_every_dir}
\alias{insert_readme_into_every_dir}
\title{This function inserts a README.md file into each dir}
\usage{
insert_readme_into_every_dir()
}
\arguments{
\item{license}{may be "mit"}
}
\value{
none
}
\description{
This function inserts a README.md file into each dir
}
\examples{
insert+readme_into_every_dir()


}
\keyword{README.md}
